fix: move futureroute code from a strange place to where it belongs
A positive side effect is that futureroutes update when landing from hsl.fi
